Move LivePreviewButton in to a file
